Fixes for re-enabling more MSVC level 4 warnings: content/renderer/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Possibly-uninitialized local variable

This also contains a style fix (wrong indentation on class section markers).

BUG=81439
TEST=none
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/372093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281637 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 176fe9ad192f383d92336af2e7721d6cf7b59a50

+++ content/renderer/pepper/pepper_video_capture_host.cc
@@ -114,7 +114,9 @@ void PepperVideoCaptureHost::PostErrorReply() {
   // the capture.
   SetStatus(PP_VIDEO_CAPTURE_STATUS_STOPPED, true);
   host()->SendUnsolicitedReply(
-      pp_resource(), PpapiPluginMsg_VideoCapture_OnError(PP_ERROR_FAILED));
+      pp_resource(),
+      PpapiPluginMsg_VideoCapture_OnError(
+          static_cast<uint32_t>(PP_ERROR_FAILED)));
 }
 
 void PepperVideoCaptureHost::OnFrameReady(
